Moved TabContent as well
