+++ src/app/pokedex/page.tsx
@@ -20,6 +20,7 @@ import { useEffect, useMemo, useState } from "react";
 import { FilterInput } from "../../components/FilterInput";
 import MoveTable from "../../components/MoveTable";
 import PokemonModal from "../../components/PokemonModal";
+import TabContent from "../../components/TabContent";
 import TypeChartCell from "../../components/TypeChartCell";
 import { ExtraTypeAbility } from "../data/abilities/ExtraTypeAbility";
 import { TypeImmunityAbility } from "../data/abilities/TypeImmunityAbility";
@@ -33,7 +34,6 @@ import { TectonicData } from "../data/tectonic/TectonicData";
 import { Tribe } from "../data/tectonic/Tribe";
 import { uniq } from "../data/util";
 import PokemonTable from "./components/PokemonTable";
-import TabContent from "./components/TabContent";
 import TableCell from "./components/TableCell";
 import TableHeader from "./components/TableHeader";
